Move kernel_info for adaptation onto KernelClient
